pyproject.toml
[project]
name = "reki-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "crewai",
    "python-dotenv",
    "requests",
    "xai-sdk",
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["reki_agent"]